+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - warp.fem: Simplified querying neighboring cell quantities when integrating on sides using new
   `warp.fem.cells()`, `warp.fem.to_inner_cell()`, `warp.fem.to_outer_cell()` operators
 - Show an error message when the type returned by a function differs from its annotation, which would have led to the compilation stage failing.
+- Clarify that `randn()` samples a normal distribution of mean 0 and variance 1.
 
 ### Fixed